requirements.txt
strands-agents
strands-tools
mcp
pydantic-settings
